refactor(dashboard): decompose fsm-hub.ts God File into 7 modules
fsm-hub.ts grew to 2145 lines across 25+ PRs, 6x over the project's
300-line split threshold. Decompose into a layered module graph while
preserving all external APIs via re-exports.

- fsm-hub-types.ts: types, constants, displayState, fmtDuration
- fsm-hub-derivations.ts: pure observation/timeline functions
- fsm-hub-diagnostics.ts: lane analysis, operational insight engine
- fsm-hub-health-panels.ts: Measurement/Invariants/Recovery cards
- fsm-hub-pipeline-panels.ts: Hero/Pipeline/Meaning/Graph panels
- fsm-hub-timeline-panels.ts: Swimlane + TransitionTrail panels
- fsm-hub.ts: orchestrator + StatusBar + SkeletonLayout + re-exports

Includes upstream changes from #7301 (skeleton), #7312 (Korean labels),
